feat(crawler): added grouped collection
Added a new argument `-g` or `--group` to the crawler, so the number of sampling points can be reduced dramatically. Instead of sampling 1004 points, now it only takes 132 points to cover the Taipei City. The groups further divide the 132 points into four sets of 33 points, which can cut down execution time from 1.5 hour to 3.5 minutes for each group.